+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added a server id option to `VisibilityAttributesPlugin` that can be used if your server is a player.
 - `ServerEventSender::send` now only sends events for connected clients plus the 'server as player'.
+- Removed `Visibility` to avoid conflicts with bevy's Visibility type, and now `VisibilityCondition` is a component.
 
 ### Added